pyproject.toml
[project]
name = "nanokaleaks"
version = "0.1.0"
description = "View newest characters from nanoka.cc"
requires-python = ">=3.11"
dependencies = [
    "customtkinter",
    "requests",
    "pillow",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nanoka_viewer"]
